+++ frontend/src/schemas/migrations.ts
@@ -1,10 +1,14 @@
 import { renderPanel } from "../api/renderPanel";
 import { runAsyncFunction } from "../components/utils";
 import { FavaExtenstionContext } from "../extension";
-import * as v1 from "./v1";
-import * as v2 from "./v2";
+import * as v1 from "./v1/v1";
+import * as dashboardv2 from "./v2/dashboard";
 
-export function migrateV1ToV2(config: v1.Config, utils: v1.Utils, extensionContext: FavaExtenstionContext): v2.Config {
+export function migrateV1ToV2(
+  config: v1.Config,
+  utils: v1.Utils,
+  extensionContext: FavaExtenstionContext,
+): dashboardv2.Config {
   return {
     ...config,
     dashboards: config.dashboards.map((dashboard) => ({
@@ -14,11 +18,15 @@ export function migrateV1ToV2(config: v1.Config, utils: v1.Utils, extensionConte
   };
 }
 
-function migratePanelToV2(v1Panel: v1.Panel, utils: v1.Utils, extensionContext: FavaExtenstionContext): v2.Panel {
+function migratePanelToV2(
+  v1Panel: v1.Panel,
+  utils: v1.Utils,
+  extensionContext: FavaExtenstionContext,
+): dashboardv2.Panel {
   return {
     ...v1Panel,
     kind: v1Panel.type === "jinja2" ? "html" : v1Panel.type,
-    spec: async (ctx) => {
+    spec: async (params) => {
       const panel: v1.Panel = await renderPanel(v1Panel);
       if (panel.type === "jinja2" && panel.template) {
         return panel.template;
@@ -26,8 +34,8 @@ function migratePanelToV2(v1Panel: v1.Panel, utils: v1.Utils, extensionContext: 
 
       const panelCtx: v1.PanelCtx = {
         ext: extensionContext,
-        ledger: ctx.ledger,
-        fava: ctx.ledger,
+        ledger: params.ledger,
+        fava: params.ledger,
         helpers: v1.helpers,
         utils,
         panel,