+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * support dynamic, typed dashboards written in TypeScript (`dashboards.tsx`) or JavaScript (`dashboards.jsx`)
 * concurrent loading of panels
 * new `react` panel kind
+* variables per dashboard and per panel
 
 ### Changed
 * updated design using Material UI